fix: change balance type to float

+++ main.go
@@ -8,7 +8,7 @@ type Account struct {
 	ID       uint64
 	Login    string
 	Password string
-	Balance  int64
+	Balance  float64
 }
 
 func main() {
